+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 ## CHANGE LOG
 
+Version 3.6.0 *(06 February 2026)*
+-------------------------------------------
+**What's new**
+* **[Android Platform]**
+  * Supports [CleverTap Android SDK v7.7.1](https://github.com/CleverTap/clevertap-android-sdk/blob/master/docs/CTCORECHANGELOG.md#version-771-december-2-2025).
+
+* **[iOS Platform]**
+  * Supports [CleverTap iOS SDK v7.4.2](https://github.com/CleverTap/clevertap-ios-sdk/blob/master/CHANGELOG.md#version-742-january-14-2026).
+
+**API changes**
+* **[Android and iOS Platform]**
+  * New API: `getVariants()` - Returns a list of A/B experiment variants for the current user.
+  * Updated API: `discardInAppNotifications({bool dismissInAppIfVisible = false})` - Added optional parameter to immediately dismiss any currently visible InApp notification in addition to discarding queued notifications.
+
 Version 3.5.3 *(22 September 2025)*
 -------------------------------------------
 **What's new**